Refactor fixed point transtormation out of CollapseTuple
